Teach usePagination to load previous data
- Use this on the constructor dashboard to enable
  someone to load "Newer puzzles"

Note: this did require the creation of a composite index. When
I loaded the page that involved the composite index, firebase
provided me with a URL that loaded a dialog within firebase
that was offering to create the index on my behalf. I'm not yet
sure if there's a programmatic way to do this in advance of
deploying this code.

Specifically, we'll need to add this index, which loads data
in a reverse of the manner that the Constructor Dashboard
currently loads data:
- `a: ASC, p: ASC, __name__: ASC`

This enables us to use the first result of the paginated query
as the argument to `endBefore` in conjunction with `limitToLast`.