Add byline to puzzle page
Adds the title and author name to the puzzle-solving interface. Includes the patron icon and adapts about as well as could be expected with absurdly long titles. Seems to adapt properly to odd aspect ratios, both of the window and puzzle. I elected to not make the author name link out directly so that mobile users don't accidentally tap it or something.